Added distutils setup.py file for installation
I added a setup.py file that can do the installation with python setup.py install.
This will also build the manpage and when specifying the -m /path/to/manpage/location will also install the manpage.
The binary is also installed.
The Makefile was modified slightly because pgloade.py was moved to scripts/pgloader.
The Makefile should still work though.