Improve summary output.
Try at having a deterministic ouput of it, which still apparently is not
always the case when using SBCL, now that it's been switched to using
the explicit *terminal-io* rather than t.

This change is needed for CCL support, though, where you don't get to
write to the same stream from different threads.