Default to a static list of PostgreSQL keywords.
In some cases (such as when using a very old PostgreSQL instance or an
Amazon Redshift service, as in #255), the function pg_get_keywords()
does not exists but we assume that pgloader might still be able to
complete its job.

We're better off with a static list of keywords than with a unhandled
error here, so let's see what happens next with Redshift.