Fix the drop indexes option again, fix #251.
The index and constraint names given by PostgreSQL catalogs should not
be second guessed, we need to just quote them. The identifier down
casing is interesting when we get identifiers from other system for a
migration, but are wrong when dropping existing indexes in PostgreSQL.

Also, the :null special value from Postmodern was routing the code
badly, just transform it manually to nil when fetching the index list,
manually.