Handle default values particularities for IXF.
It turns out that IXF format might embed particulars of the systems the
data comes from (Informix or DB2) and we need to process some strings so
that they are compatible with PostgreSQL ("CURRENT TIMESTAMP" here).

See #272 that should be fixed here, but being in the blind, don't close
the issue just yet.