Fix MS SQL int conversion, fix #282.
This is a blind fix that I hope is the one needed, wherein -1 get sent
to unsigned-to-sign which fails to handle it as expected. The protocol
and driver are then sending unsigned ints and that's what we now handle
in the CFFI layer.

This allows to push the attempt to ease testing of it, and allows me
also to cut a release without modified files handling around.