Trick the reporting to show non-zero timings.
When calling lparallel:receive-results from the main threads we loose
the ability to measure proper per-table processing times, because it all
happens in parallel and the main threads seems to receive events in the
same millisecond as when the worker is started, meaning it's all 0.0s.

So when we don't have "secs" stats, pick the greatest of read or write
time, which we do have from the worker threads themselves.

The number are still wrong, but less so than the "0.0s" displayed before
this patch.