Travis: prepare an "err" schema.
The test/errors.load set the search_path to include the 'err' schema,
which is to be created by the test itself. PostgreSQL 9.1 raises an
error where 9.4 and following just accept the setting, and Travis runs a
9.1 PostgreSQL.

Let's just create the schema before-hand so that we can still run tests
against SET search_path from the load file.